+++ src/main/java/com/teliolabs/corba/data/dto/AdditionalInfo.java
@@ -8,8 +8,8 @@ import lombok.Setter;
 @Setter
 public class AdditionalInfo {
     private final PathType pathType;
-    private final byte tupleA;
-    private final byte tupleB;
+    private final short tupleA;
+    private final short tupleB;
     private final String tuple;
 
     public AdditionalInfo(String orderNumber) {
@@ -18,12 +18,12 @@ public class AdditionalInfo {
 
         // Check if -1 values are at the start or end and assign remaining numbers
         if (parts[0].equals("-1") && parts[1].equals("-1")) {
-            tupleA = Byte.parseByte(parts[2]);
-            tupleB = Byte.parseByte(parts[3]);
+            tupleA = Short.parseShort(parts[2]);
+            tupleB = Short.parseShort(parts[3]);
             pathType = PathType.PROTECTION;
         } else if (parts[parts.length - 1].equals("-1") && parts[parts.length - 2].equals("-1")) {
-            tupleA = Byte.parseByte(parts[0]);
-            tupleB = Byte.parseByte(parts[1]);
+            tupleA = Short.parseShort(parts[0]);
+            tupleB = Short.parseShort(parts[1]);
             pathType = PathType.MAIN;
         } else {
             tupleA = -1;
